General refactor

+++ src/main/java/br/com/laranja/springcrud/domain/service/impl/EventoServiceImpl.java
@@ -1,15 +1,10 @@
 package br.com.laranja.springcrud.domain.service.impl;
 
 
-import br.com.laranja.springcrud.domain.dto.EventoRequest;
-import br.com.laranja.springcrud.domain.model.Evento;
-import br.com.laranja.springcrud.domain.model.Tela;
-import br.com.laranja.springcrud.domain.model.TipoEvento;
+import br.com.laranja.springcrud.domain.dto.evento.EventoRequest;
+import br.com.laranja.springcrud.domain.model.*;
 import br.com.laranja.springcrud.domain.service.EventoService;
-import br.com.laranja.springcrud.infrastructure.exception.EventoNotFoundException;
-import br.com.laranja.springcrud.infrastructure.exception.TelaNotFoundException;
-import br.com.laranja.springcrud.infrastructure.exception.TipoEventoNotFoundException;
-import br.com.laranja.springcrud.infrastructure.exception.VersaoNotFoundException;
+import br.com.laranja.springcrud.infrastructure.exception.*;
 import br.com.laranja.springcrud.infrastructure.repository.EventoRepository;
 import br.com.laranja.springcrud.infrastructure.repository.TelaRepository;
 
@@ -29,8 +24,7 @@ public class EventoServiceImpl implements EventoService {
     private final TipoEventoRepository tipoEventoRepository;
 
     @Override
-    public List<Evento> getAllEventos() {
-        return eventoRepository.findAll();
+    public List<Evento> getAllEventos() {return eventoRepository.findAll();
     }
 
     @Override
@@ -101,8 +95,11 @@ public class EventoServiceImpl implements EventoService {
 
     @Transactional
     @Override
-    public void deleteEventoById(Long idEvento) {
-
+    public void deleteEventoById(Long idEvento) throws EventoNotFoundException {
+        Optional<Evento> evento = Optional.ofNullable(this.getEventoById(idEvento));
+        if (!evento.isPresent()){
+            throw new EventoNotFoundException(idEvento);
+        }
         eventoRepository.deleteByIdEvento(idEvento);
     }
 }
