+++ src/main/java/br/com/laranja/springcrud/domain/service/impl/ProjetoServiceImpl.java
@@ -1,14 +1,13 @@
 package br.com.laranja.springcrud.domain.service.impl;
 
-import br.com.laranja.springcrud.domain.dto.ProjetoRequest;
+import br.com.laranja.springcrud.domain.dto.projeto.ProjetoRequest;
 import br.com.laranja.springcrud.domain.model.Projeto;
-import br.com.laranja.springcrud.domain.model.Usuario;
-import br.com.laranja.springcrud.domain.model.Versao;
 import br.com.laranja.springcrud.domain.service.ProjetoService;
 import br.com.laranja.springcrud.infrastructure.exception.ProjetoNotFoundException;
-import br.com.laranja.springcrud.infrastructure.exception.UsuarioNotFoundException;
 import br.com.laranja.springcrud.infrastructure.repository.ProjetoRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -20,17 +19,21 @@ import java.util.Optional;
 public class ProjetoServiceImpl implements ProjetoService {
     private  final ProjetoRepository projetoRepository;
 
+
     @Override
+    @Cacheable(value = "projetos")
     public List<Projeto> getAllProjetos() {
         return projetoRepository.findAll();
     }
 
     @Override
+    @CacheEvict(value = "projetos", allEntries = true)
     public Projeto getProjetoById(Long idProjeto) {
         return projetoRepository.findById(idProjeto).orElseThrow( () -> new ProjetoNotFoundException(idProjeto));
     }
 
     @Override
+    @CacheEvict(value = "projetos", allEntries = true)
     public Projeto createProjeto(ProjetoRequest projetoRequest) {
         Projeto projeto = Projeto.builder()
                 .nome(projetoRequest.getNome())
@@ -40,6 +43,7 @@ public class ProjetoServiceImpl implements ProjetoService {
     }
 
     @Override
+    @CacheEvict(value = "projetos", allEntries = true)
     public Projeto updateProjetoById(Long idProjeto, ProjetoRequest projetoRequest) {
         Optional<Projeto> projetoOptional = projetoRepository.findById(idProjeto);
         //verifica se tem o id no banco
@@ -57,7 +61,12 @@ public class ProjetoServiceImpl implements ProjetoService {
 
     @Transactional
     @Override
+    @CacheEvict(value = "projetos", allEntries = true)
     public void deleteProjetoById(Long idProjeto) {
+        Optional<Projeto> projeto = Optional.ofNullable(this.getProjetoById(idProjeto));
+        if (!projeto.isPresent()) {
+            throw new ProjetoNotFoundException(idProjeto);
+        }
         projetoRepository.deleteByIdProjeto(idProjeto);
     }
 }
