+++ src/main/java/br/com/laranja/springcrud/domain/service/impl/RequisicaoServiceImpl.java
@@ -1,11 +1,13 @@
 package br.com.laranja.springcrud.domain.service.impl;
 
-import br.com.laranja.springcrud.domain.dto.RequisicaoRequest;
+import br.com.laranja.springcrud.domain.dto.requiscao.RequisicaoRequest;
 import br.com.laranja.springcrud.domain.model.Evento;
 import br.com.laranja.springcrud.domain.model.Requisicao;
+import br.com.laranja.springcrud.domain.model.Tela;
 import br.com.laranja.springcrud.domain.service.RequisicaoService;
 import br.com.laranja.springcrud.infrastructure.exception.EventoNotFoundException;
 import br.com.laranja.springcrud.infrastructure.exception.RequisicaoNotFoundException;
+import br.com.laranja.springcrud.infrastructure.exception.TelaNotFoundException;
 import br.com.laranja.springcrud.infrastructure.repository.EventoRepository;
 import br.com.laranja.springcrud.infrastructure.repository.RequisicaoRepository;
 import lombok.RequiredArgsConstructor;
@@ -82,6 +84,11 @@ public class RequisicaoServiceImpl implements RequisicaoService {
     @Transactional
     @Override
     public void deleteRequisicaoById(Long idRequisicao) {
+        Optional<Requisicao> requisicao = Optional.ofNullable(this.getRequisicaoById(idRequisicao));
+        if (!requisicao.isPresent()) {
+            throw new RequisicaoNotFoundException(idRequisicao);
+        }
+
         requisicaoRepository.deleteByIdRequisicao(idRequisicao);
     }
