+++ src/main/java/br/com/laranja/springcrud/domain/service/impl/TelaServiceImpl.java
@@ -1,12 +1,13 @@
 package br.com.laranja.springcrud.domain.service.impl;
 
 
-import br.com.laranja.springcrud.domain.dto.TelaRequest;
+import br.com.laranja.springcrud.domain.dto.tela.TelaRequest;
+import br.com.laranja.springcrud.domain.model.Evento;
 import br.com.laranja.springcrud.domain.model.Projeto;
 import br.com.laranja.springcrud.domain.model.Tela;
 import br.com.laranja.springcrud.domain.model.Versao;
 import br.com.laranja.springcrud.domain.service.TelaService;
-import br.com.laranja.springcrud.infrastructure.exception.ProjetoNotFoundException;
+import br.com.laranja.springcrud.infrastructure.exception.EntityWithDependentsException;
 import br.com.laranja.springcrud.infrastructure.exception.TelaNotFoundException;
 import br.com.laranja.springcrud.infrastructure.exception.VersaoNotFoundException;
 import br.com.laranja.springcrud.infrastructure.repository.TelaRepository;
@@ -61,10 +62,14 @@ public class TelaServiceImpl implements TelaService {
         Optional<Tela> TelaOptional = telaRepository.findById(idTela);
 
         if (!TelaOptional.isPresent()) {
-            throw new VersaoNotFoundException(idTela);
+            throw new TelaNotFoundException(idTela);
         }
 
         Optional<Versao> Optionalversao = versaoRepository.findById(telaRequest.getIdVersao());
+
+        if (Optionalversao.isEmpty() ){
+            throw  new VersaoNotFoundException(telaRequest.getIdVersao());
+        }
 
         Versao versaoExistent = Optionalversao.get();
 
@@ -85,6 +90,10 @@ public class TelaServiceImpl implements TelaService {
     @Transactional
     @Override
     public void deleteTelaById(Long idTela) {
-       telaRepository.deleteByIdTela(idTela);
+        Optional<Tela> tela = Optional.ofNullable(this.getTelaById(idTela));
+        if (!tela.isPresent()){
+            throw new TelaNotFoundException(idTela);
+        }
+        telaRepository.deleteByIdTela(idTela);
     }
 }
