+++ src/main/java/br/com/laranja/springcrud/domain/service/impl/TipoEventoServiceImpl.java
@@ -1,21 +1,25 @@
 package br.com.laranja.springcrud.domain.service.impl;
 
 
-import br.com.laranja.springcrud.domain.dto.TipoEventoRequest;
+import br.com.laranja.springcrud.domain.dto.tipoEvento.TipoEventoRequest;
+import br.com.laranja.springcrud.domain.dto.tipoEvento.TipoEventoResponse;
+import br.com.laranja.springcrud.domain.model.Evento;
 import br.com.laranja.springcrud.domain.model.Projeto;
 import br.com.laranja.springcrud.domain.model.TipoEvento;
-import br.com.laranja.springcrud.domain.model.Usuario;
+import br.com.laranja.springcrud.domain.model.Versao;
 import br.com.laranja.springcrud.domain.service.TipoEventoService;
+import br.com.laranja.springcrud.infrastructure.exception.EntityWithDependentsException;
 import br.com.laranja.springcrud.infrastructure.exception.EventoNotFoundException;
-import br.com.laranja.springcrud.infrastructure.exception.ProjetoNotFoundException;
 import br.com.laranja.springcrud.infrastructure.exception.TipoEventoNotFoundException;
 import br.com.laranja.springcrud.infrastructure.exception.VersaoNotFoundException;
+import br.com.laranja.springcrud.infrastructure.repository.EventoRepository;
 import br.com.laranja.springcrud.infrastructure.repository.TipoEventoRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,18 +27,22 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class TipoEventoServiceImpl implements TipoEventoService {
     private final TipoEventoRepository tipoeventoRepository;
+    private final EventoRepository eventoRepository;
 
     @Override
+    @Cacheable(value = "tipoEvento")
     public List<TipoEvento> getAllTipoEventos() {
         return tipoeventoRepository.findAll();
     }
 
     @Override
+    @CacheEvict(value = "tipoEvento", allEntries = true)
     public TipoEvento getTipoEventoById(Long idTipoEvento) {
         return tipoeventoRepository.findById(idTipoEvento).orElseThrow(() -> new EventoNotFoundException(idTipoEvento));
     }
 
     @Override
+    @CacheEvict(value = "tipoEvento", allEntries = true)
     public TipoEvento createTipoEvento(TipoEventoRequest tipoEventoRequest) {
         TipoEvento tipoevento = TipoEvento.builder()
                 .situacao(tipoEventoRequest.isSituacao())
@@ -45,6 +53,7 @@ public class TipoEventoServiceImpl implements TipoEventoService {
     }
 
     @Override
+    @CacheEvict(value = "tipoEvento", allEntries = true)
     public TipoEvento updateTipoEventoById(Long idTipoEvento, TipoEventoRequest tipoEventoRequest) {
         Optional<TipoEvento> TipoEventoOpcional = tipoeventoRepository.findById(idTipoEvento);
 
@@ -63,7 +72,13 @@ public class TipoEventoServiceImpl implements TipoEventoService {
 
     @Transactional
     @Override
-    public void deleteTipoEventoById(Long id) {
+    @CacheEvict(value = "tipoEvento", allEntries = true)
+    public void deleteTipoEventoById(Long id) throws RuntimeException {
+        Optional<TipoEvento> tipoEventoOptional = Optional.ofNullable(this.getTipoEventoById(id));
+        if (!tipoEventoOptional.isPresent()){
+            throw new TipoEventoNotFoundException(id);
+        }
+
         tipoeventoRepository.deleteByIdTipoEvento(id);
     }
 }
